Consolidate tana module layout: merge core into nodes, move loader to graph
- Merge get_tuple_value from query/core.py into query/nodes.py (alongside
  get_field_values), delete core.py
- Move load_workspace from io/json.py to graph/loader.py, delete io/
- Rename graph/workspace.py to graph/graph.py to match class name TanaGraph
  (resolves workspace-vs-graph naming confusion with tana/workspace.py facade)
- Update all imports and BUILD.bazel targets
