+++ tana/domain/BUILD.bazel
@@ -14,7 +14,6 @@ py_library(
     srcs = ["nodes.py"],
     visibility = ["//:__subpackages__"],
     deps = [
-        ":constants",
         ":types",
         "@pypi//pydantic",
     ],
