+++ tana/export/BUILD.bazel
@@ -17,10 +17,9 @@ py_binary(
         "//tana/domain:constants",
         "//tana/domain:nodes",
         "//tana/domain:types",
-        "//tana/graph:workspace",
+        "//tana/graph",
+        "//tana/graph:loader",
         "//tana/graph:wrappers",
-        "//tana/io:json",
-        "//tana/query:core",
         "//tana/query:nodes",
         "//tana/render:html_utils",
         "//tana/render:inline_refs",
@@ -36,7 +35,7 @@ py_binary(
         "//tana/domain:constants",
         "//tana/domain:nodes",
         "//tana/domain:types",
-        "//tana/graph:workspace",
+        "//tana/graph",
         "//tana/graph:wrappers",
     ],
 )
@@ -48,7 +47,7 @@ py_binary(
     deps = [
         "//tana:workspace",
         "//tana/domain:types",
-        "//tana/graph:workspace",
+        "//tana/graph",
         "//tana/query/search:materializer",
         "//tana/query/search:parser",
     ],
