+++ tana/export/convert.py
@@ -17,11 +17,10 @@ from tana.domain.constants import (
 )
 from tana.domain.nodes import BaseNode, CodeBlockNode, TupleNode, VisualNode
 from tana.domain.types import NodeId
-from tana.graph.workspace import TanaGraph
+from tana.graph.graph import TanaGraph
+from tana.graph.loader import load_workspace
 from tana.graph.wrappers import is_wrapper
-from tana.io.json import load_workspace
-from tana.query.core import get_tuple_value
-from tana.query.nodes import get_image_url
+from tana.query.nodes import get_image_url, get_tuple_value
 from tana.render.html_utils import DATE_SPAN_PATTERN, NODE_SPAN_PATTERN, find_inline_node_refs, html_to_markdown
 from tana.render.inline_refs import parse_inline_date
