+++ tana/query/nodes.py
@@ -1,11 +1,48 @@
 from __future__ import annotations
 
 from collections.abc import Iterator
+from typing import TYPE_CHECKING
 
 from tana.domain.constants import MEDIA_KEY_ID, MIN_TUPLE_CHILDREN
 from tana.domain.nodes import BaseNode, TupleNode
-from tana.graph.workspace import TanaGraph
-from tana.query.core import get_tuple_value
+from tana.domain.types import NodeId
+
+if TYPE_CHECKING:
+    from tana.graph.graph import TanaGraph
+
+
+def get_tuple_value(node: BaseNode, key: NodeId | str, store: TanaGraph) -> BaseNode | None:
+    """Return the first value node from a tuple keyed by `key`.
+
+    Supports two shapes:
+    - node is a tuple node: children[0] is the key id, children[1:] are values
+    - node is a container: search its child tuple nodes for one where
+      tuple.children[0] == key, then return the second child
+    """
+    key_str = str(key)
+
+    children = list(node.children)
+
+    # Case 1: node itself is a tuple — check its key
+    if children:
+        first = children[0]
+        if str(first) == key_str and len(children) >= 2:
+            return store.get(children[1])
+
+    # Case 2: search child tuples under this node
+    for cid in children:
+        try:
+            t = store[cid]
+        except KeyError:
+            continue
+        t_children = list(t.children)
+        if not t_children:
+            continue
+        if str(t_children[0]) != key_str:
+            continue
+        if len(t_children) >= 2:
+            return store.get(t_children[1])
+    return None
 
 
 def get_field_values(node: BaseNode, field_name: str, store: TanaGraph) -> Iterator[str]:
