+++ tana/query/search/BUILD.bazel
@@ -11,7 +11,7 @@ py_library(
         "//tana/domain:nodes",
         "//tana/domain:search",
         "//tana/domain:types",
-        "//tana/graph:workspace",
+        "//tana/graph",
         "//tana/query:filters",
     ],
 )
@@ -25,7 +25,7 @@ py_library(
         ":parser",
         "//tana/domain:nodes",
         "//tana/domain:types",
-        "//tana/graph:workspace",
+        "//tana/graph",
     ],
 )
 
@@ -38,6 +38,6 @@ py_library(
         "//tana/domain:nodes",
         "//tana/domain:search",
         "//tana/domain:types",
-        "//tana/graph:workspace",
+        "//tana/graph",
     ],
 )
